refactor(memory): use unsigned long long for memory size
- Changed total_kb and available_kb types from long to unsigned long
  long
- Updated sscanf format specifiers from %ld to %llu

+++ include/memory.h
@@ -1,3 +1,3 @@
-long GetTotMem(void);
-long GetAvailMem(void);
+unsigned long long GetTotMem(void);
+unsigned long long GetAvailMem(void);
 double GetMemUsage(void);
